move focalheatmaploss to losses module
