+++ mmpose/models/heads/cid_head.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from ..builder import HEADS
+from ..builder import HEADS, build_loss
 
 
 def _sigmoid(x):
@@ -58,7 +58,7 @@ class CIDHead(nn.Module):
         self.heatmap_conv = nn.Conv2d(gfd_channels, num_joints, 1, 1, 0)
 
         # loss
-        self.heatmap_loss = FocalLoss()
+        self.heatmap_loss = build_loss(dict(type='FocalHeatmapLoss'))
         self.contrastive_loss = ContrastiveLoss()
 
         # initialize
@@ -329,48 +329,6 @@ class SpatialAtten(nn.Module):
         input_feats = torch.cat((input_feats, relative_coords), dim=1)
         mask = self.conv(input_feats).sigmoid()
         return global_features * mask
-
-
-class FocalLoss(nn.Module):
-
-    def __init__(self, alpha=2, beta=4):
-        super(FocalLoss, self).__init__()
-        self.alpha = alpha
-        self.beta = beta
-
-    def forward(self, pred, gt, mask=None):
-        """Modified focal loss.
-
-        Exactly the same as CornerNet.
-        Runs faster and costs a little bit more memory
-        Arguments:
-          pred (batch x c x h x w)
-          gt_regr (batch x c x h x w)
-        """
-        pos_inds = gt.eq(1).float()
-        neg_inds = gt.lt(1).float()
-
-        if mask is not None:
-            pos_inds = pos_inds * mask
-            neg_inds = neg_inds * mask
-
-        neg_weights = torch.pow(1 - gt, self.beta)
-
-        loss = 0
-
-        pos_loss = torch.log(pred) * torch.pow(1 - pred, self.alpha) * pos_inds
-        neg_loss = torch.log(1 - pred) * torch.pow(
-            pred, self.alpha) * neg_weights * neg_inds
-
-        num_pos = pos_inds.float().sum()
-        pos_loss = pos_loss.sum()
-        neg_loss = neg_loss.sum()
-
-        if num_pos == 0:
-            loss = loss - neg_loss
-        else:
-            loss = loss - (pos_loss + neg_loss) / num_pos
-        return loss
 
 
 class ContrastiveLoss(nn.Module):
