+++ mmpose/models/losses/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .classfication_loss import BCELoss
-from .heatmap_loss import AdaptiveWingLoss
+from .heatmap_loss import AdaptiveWingLoss, FocalHeatmapLoss
 from .mesh_loss import GANLoss, MeshLoss
 from .mse_loss import JointsMSELoss, JointsOHKMMSELoss
 from .multi_loss_factory import AELoss, HeatmapLoss, MultiLossFactory
@@ -9,8 +9,23 @@ from .regression_loss import (BoneLoss, L1Loss, MPJPELoss, MSELoss, RLELoss,
                               WingLoss)
 
 __all__ = [
-    'JointsMSELoss', 'JointsOHKMMSELoss', 'HeatmapLoss', 'AELoss',
-    'MultiLossFactory', 'MeshLoss', 'GANLoss', 'SmoothL1Loss', 'WingLoss',
-    'MPJPELoss', 'MSELoss', 'L1Loss', 'BCELoss', 'BoneLoss',
-    'SemiSupervisionLoss', 'SoftWingLoss', 'AdaptiveWingLoss', 'RLELoss'
+    'JointsMSELoss',
+    'JointsOHKMMSELoss',
+    'HeatmapLoss',
+    'AELoss',
+    'MultiLossFactory',
+    'MeshLoss',
+    'GANLoss',
+    'SmoothL1Loss',
+    'WingLoss',
+    'MPJPELoss',
+    'MSELoss',
+    'L1Loss',
+    'BCELoss',
+    'BoneLoss',
+    'SemiSupervisionLoss',
+    'SoftWingLoss',
+    'AdaptiveWingLoss',
+    'RLELoss',
+    'FocalHeatmapLoss',
 ]
